IdCSC, suporte a números maiores que 32767 do Int16.

+++ Shared.NFe.Utils/InformacoesSuplementares/ExtinfNFeSupl.cs
@@ -474,7 +474,7 @@ namespace NFe.Utils.InformacoesSuplementares
             var ambiente = (int)nfe.infNFe.ide.tpAmb;
 
             //Identificador do CSC (Código de Segurança do Contribuinte no Banco de Dados da SEFAZ). Informar sem os zeros não significativos
-            var idCsc = Convert.ToInt16(cIdToken);
+            var idCsc = Convert.ToInt32(cIdToken);
 
             string dadosBase;
